DFPL-1201: move submitted and postSubmit callbacks to new event
