+++ ccd-definition/CaseEvent/CareSupervision/GatekeepingListing.json
@@ -15,5 +15,15 @@
     "CallBackURLAboutToStartEvent": "${CCD_DEF_CASE_SERVICE_BASE_URL}/callback/list-gatekeeping-hearing/about-to-start",
     "CallBackURLAboutToSubmitEvent": "${CCD_DEF_CASE_SERVICE_BASE_URL}/callback/list-gatekeeping-hearing/about-to-submit",
     "CallBackURLSubmittedEvent": "${CCD_DEF_CASE_SERVICE_BASE_URL}/callback/list-gatekeeping-hearing/submitted"
+  },
+  {
+    "LiveFrom": "01/01/2017",
+    "CaseTypeID": "CARE_SUPERVISION_EPO",
+    "ID": "internal-change-add-gatekeeping",
+    "Name": "Post-process gatekeeping order",
+    "PreConditionState(s)": "*",
+    "PostConditionState": "*",
+    "SecurityClassification": "Public",
+    "CallBackURLAboutToSubmitEvent": "${CCD_DEF_CASE_SERVICE_BASE_URL}/callback/list-gatekeeping-hearing/post-submit-callback/about-to-submit"
   }
 ]
