+++ ccd-definition/CaseEvent/CareSupervision/MultiState.json
@@ -59,16 +59,6 @@
   {
     "LiveFrom": "01/01/2017",
     "CaseTypeID": "CARE_SUPERVISION_EPO",
-    "ID": "internal-change-add-gatekeeping",
-    "Name": "Post-process gatekeeping order",
-    "PreConditionState(s)": "*",
-    "PostConditionState": "*",
-    "SecurityClassification": "Public",
-    "CallBackURLAboutToSubmitEvent": "${CCD_DEF_CASE_SERVICE_BASE_URL}/callback/add-gatekeeping-order/post-submit-callback/about-to-submit"
-  },
-  {
-    "LiveFrom": "01/01/2017",
-    "CaseTypeID": "CARE_SUPERVISION_EPO",
     "ID": "internal-change-placement",
     "Name": "Post-process placement",
     "PreConditionState(s)": "*",
